requirements.txt
Flask
Flask-SQLAlchemy
flask-login
Flask-WTF
Flask-Cors
python-dotenv
email-validator